wired the beans and made the service return stubbed data for now

+++ QuestionBankApp/src/main/java/com/java/applications/questionbank/entities/Question.java
@@ -12,5 +12,7 @@ public class Question {
 
 	@Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private long id;
+    private int id;
+	
+	
 }
